refactor(jest-config): rename jest.config.js to jest-monorepo.config.js
