+++ packages/shared/jest-config/lib/index.js
@@ -1,5 +1,5 @@
 const ModuleMapper = require('./ModuleMapper')
-const monorepo = require('./jest.config')
+const monorepo = require('./jest-monorepo.config')
 const base = require('./jest-base.config')
 
 module.exports = {
