+++ packages/shared/jest-config/lib/index.test.js
@@ -1,6 +1,6 @@
 const { deepStrictEqual, throws } = require('assert')
 const ModuleMapper = require('./ModuleMapper')
-const monorepo = require('./jest.config')
+const monorepo = require('./jest-monorepo.config')
 const base = require('./jest-base.config')
 
 {
