+++ packages/shared/jest-config/lib/jest-monorepo.config.js
